+++ meteor-app/imports/api/users/server/publications.js
@@ -43,21 +43,26 @@ Meteor.methods({
   },
   updateMemberPassword(formData, confirmPass) {
     const userId = formData.userId
-    const oldEmail = Meteor.user().emails
-    try {
-      // update member form
-      Meteor.call('members.update', formData._id, formData)
-      // update user password
-      Accounts.setPassword(userId, confirmPass, { logout: false })
-      // update user email
-      Accounts.setUsername(userId, formData.email)
-      if (oldEmail) {
-        Accounts.removeEmail(userId, oldEmail[0].address)
-        Accounts.addEmail(userId, formData.email)
+    if (Meteor.user()) {
+      const oldEmail = Meteor.user().emails
+      try {
+        // update member form
+        Meteor.call('members.update', formData._id, formData)
+
+        // update user password
+        Accounts.setPassword(userId, confirmPass, { logout: false })
+        // update user email
+        Accounts.setUsername(userId, formData.email)
+        if (oldEmail) {
+          Accounts.removeEmail(userId, oldEmail[0].address)
+          Accounts.addEmail(userId, formData.email)
+        }
+        return { status: 'success', message: `Updated user` }
+      } catch (e) {
+        return { status: 'failed', message: `Error updating user: ${e.message}` }
       }
-      return { status: 'success', message: `Updated user` }
-    } catch (e) {
-      return { status: 'failed', message: `Error updating user: ${e.message}` }
+    } else {
+      return { status: 'success', message: `No user associated with account` }
     }
   },
   setPassword({ id, newPassword }, logout = true) {
@@ -76,5 +81,5 @@ Meteor.methods({
     } catch (e) {
       return { status: 'failed', message: `Error sending password reset: ${e.message}` }
     }
-  }
+  },
 })