+++ meteor-app/tests/cypress/integration/pa/e2e-multi-add-tough.js
@@ -54,7 +54,7 @@ describe('Login into Kiosk', function() {
     cy.visit('/admin/userprofiles/')
     loginAsAdmin()
 
-    cy.get('div[class="header"]')
+    cy.get('td')
       .contains(member.name)
       .should('exist')
       .click()