+++ meteor-app/CHANGELOG.md
@@ -8,6 +8,9 @@
 - Added view of charges with 'refresh' button
 - Add reconciliation to charges and carts, improved layouts
 - Add admin button to edit user details
+- Add b2b member portal
+- Fix coding problems with certain test scripts
+- Add new service test scripts for input and UI
 - Refactor factories code
 - Added member.paymentEmails array
 - Fix email to be lower case